Yet more improvements

+++ src/api/blockchain/confirm_transaction_params.rs
@@ -2,12 +2,21 @@
   // Probably little will be gained boxing a variant of 128 bits
   variant_size_differences
 )]
-#[derive(Debug, Eq, PartialEq)]
+/// Transactions usually take some time to be confirmed.
+#[derive(Debug)]
 pub enum ConfirmTransactionOptions {
   /// Keeps trying fetching a transaction until a provided `number` of iteration is reached or the
   /// transaction is confirmed.
-  Tries { number: u32 },
+  Tries {
+    /// Max of 65_535 requests
+    number: u16,
+  },
   /// Keeps trying fetching a transaction until a provided `number` of iteration is reached or the
   /// transaction is confirmed. Each iteration awaits the provided `interval`.
-  TriesWithInterval { interval: core::time::Duration, number: u32 },
+  TriesWithInterval {
+    /// Any measure of time
+    interval: core::time::Duration,
+    /// Max of 65_535 requests
+    number: u16,
+  },
 }
